KAA-1096: Remove 'Download configuration' button from Profile filter in Admin UI
