+++ server/node/src/main/java/org/kaaproject/kaa/server/admin/client/mvp/view/struct/ConfigFormStructView.java
@@ -16,7 +16,12 @@
 
 package org.kaaproject.kaa.server.admin.client.mvp.view.struct;
 
+import com.google.gwt.event.dom.client.ClickEvent;
+import com.google.gwt.event.dom.client.ClickHandler;
+import com.google.gwt.user.client.Window;
+import com.google.gwt.user.client.ui.Button;
 import org.kaaproject.avro.ui.shared.RecordField;
+import org.kaaproject.kaa.common.dto.UpdateStatus;
 import org.kaaproject.kaa.server.admin.client.mvp.view.widget.RecordPanel;
 import org.kaaproject.kaa.server.admin.client.util.HasErrorMessage;
 import org.kaaproject.kaa.server.admin.client.util.Utils;
@@ -28,6 +33,8 @@ import com.google.gwt.event.shared.HandlerRegistration;
 import com.google.gwt.user.client.ui.HasValue;
 
 public class ConfigFormStructView extends BaseStructView<ConfigurationRecordFormDto, RecordField> {
+
+    private Button downloadConfigurationButton;
 
     public ConfigFormStructView(HasErrorMessage hasErrorMessage) {
         super(hasErrorMessage);
@@ -64,6 +71,38 @@ public class ConfigFormStructView extends BaseStructView<ConfigurationRecordForm
                 fireChanged();
             }
         });
+    }
+
+    @Override
+    protected void init() {
+        super.init();
+        downloadConfigurationButton = new Button(Utils.constants.downloadConfiguration());
+
+        buttonsPanel.add(downloadConfigurationButton);
+
+        downloadConfigurationButton.addClickHandler(new ClickHandler() {
+            @Override
+            public void onClick(ClickEvent clickEvent) {
+                String url = Window.Location.getHref();
+                Window.open("/kaaAdmin/servlet/kaaConfigurationDownloadServlet?schemaId=" + Utils.getSchemaIdFromUrl(url) +
+                                "&endGroupId=" + Utils.getEndpointGroupIdFromUrl(url),
+                        "_blank", "status=0,toolbar=0,menubar=0,location=0");
+            }
+        });
+    }
+
+    @Override
+    public void reset() {
+        super.reset();
+        downloadConfigurationButton.setVisible(false);
+    }
+
+    @Override
+    public void setData(ConfigurationRecordFormDto struct) {
+        super.setData(struct);
+        if (struct.getStatus()==UpdateStatus.ACTIVE) {
+            downloadConfigurationButton.setVisible(true);
+        }
     }
 
     @Override
