+++ server/node/src/main/resources/org/kaaproject/kaa/server/admin/client/i18n/KaaAdminConstants.properties
@@ -171,7 +171,7 @@ deactivate = Deactivate
 
 deactivatedBy = Deactivated by
 
-downloadConfiguration = Download configuration
+downloadConfiguration = Download
 
 deep = Deep
